+++ csrc/include/natten/cuda/fmha_hopper/kernel/fmha_kernel_builder.hpp
@@ -42,6 +42,7 @@
 namespace cutlass::fmha::kernel {
 
 template <
+    class ProblemShape,
     class Element_,
     class ElementAccumulatorQK_,
     class ElementAccumulatorPV_,
@@ -55,12 +56,14 @@ template <
 struct FmhaBuilder;
 
 template <
+    class ProblemShape,
     class Element,
     class ElementAccumulator,
     class TileShape, // BlockQO, BlockKV, BlockHead
     class Fusion,
     class... Options>
 struct FmhaBuilder<
+    ProblemShape,
     Element,
     ElementAccumulator,
     ElementAccumulator,
@@ -83,11 +86,15 @@ struct FmhaBuilder<
       ElementAccumulator,
       typename CollectiveMainloop::TileShapePV>;
 
-  using Kernel = cutlass::fmha::kernel::
-      FmhaKernelTma<CollectiveMainloop, CollectiveEpilogue, Options...>;
+  using Kernel = cutlass::fmha::kernel::FmhaKernelTma<
+      ProblemShape,
+      CollectiveMainloop,
+      CollectiveEpilogue,
+      Options...>;
 };
 
 template <
+    class ProblemShape,
     class Element,
     class ElementAccumulatorQK,
     class ElementAccumulatorPV,
@@ -98,6 +105,7 @@ template <
     class Fusion,
     class... Options>
 struct FmhaBuilder<
+    ProblemShape,
     Element,
     ElementAccumulatorQK,
     ElementAccumulatorPV,
@@ -133,6 +141,7 @@ struct FmhaBuilder<
       cutlass::fmha::kernel::IndividualTileScheduler>;
 
   using Kernel = cutlass::fmha::kernel::FmhaKernelTmaWarpSpecialized<
+      ProblemShape,
       CollectiveMainloop,
       CollectiveEpilogue,
       TileScheduler,
@@ -140,6 +149,7 @@ struct FmhaBuilder<
 };
 
 template <
+    class ProblemShape,
     class Element,
     class ElementAccumulatorQK,
     class ElementAccumulatorPV,
@@ -150,6 +160,7 @@ template <
     class Fusion,
     class... Options>
 struct FmhaBuilder<
+    ProblemShape,
     Element,
     ElementAccumulatorQK,
     ElementAccumulatorPV,
@@ -161,6 +172,7 @@ struct FmhaBuilder<
     cutlass::gemm::KernelTmaWarpSpecializedPingpong,
     Options...> {
   using Kernel = typename FmhaBuilder<
+      ProblemShape,
       Element,
       ElementAccumulatorQK,
       ElementAccumulatorPV,